requirements.txt
web3==6.11.1
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0
